+++ src/redturtle/prenotazioni/content/prenotazioni_folder.py
@@ -632,7 +632,7 @@ class IPrenotazioniFolder(model.Schema):
             default="The number of simultaneous bookings allowed for the same user.",
         ),
         required=False,
-        default=2,
+        default=0,
     )
 
     model.fieldset(