+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-2.1.2 (unreleased)
+2.1.2 (2023-10-13)
 ------------------
 
 - Add hidden booking types for operator use.